yolov5/requirements.txt
Cython
matplotlib>=3.2.2
seaborn
numpy>=1.18.5
pandas
opencv-python>=4.1.2
Pillow
PyYAML>=5.3.1
scipy>=1.4.1
tensorboard>=2.2
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0